Animation succesfully migrated to Plot_Page file
