+++ Plot_Page.py
@@ -2,6 +2,18 @@ import tkinter as tk
 from tkinter import messagebox
 from tkinter import ttk
 
+# File system access library
+import glob, os
+
+# Math functions library
+import numpy as np
+#-----------------------------------------------------------------#
+#-----------------------------------------------------------------#
+# Plot Page Code
+import tkinter as tk
+from tkinter import messagebox
+from tkinter import ttk
+
 import numpy as np
 
 # Import plotting libraries
@@ -15,7 +27,31 @@ from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg, NavigationToolb
 from matplotlib.figure import Figure
 matplotlib.use('TkAgg')
 
+f = Figure(figsize=(5,5), dpi=100)
+a = f.add_subplot(111)
+
+from sys import argv
+
 # Plot Page
+def animate(i):
+        filepath = "/home/pi/ava/"
+        os.chdir(filepath)
+        
+        pullData = open('DataPlotFile.txt','r').read()
+        dataList = pullData.split('\n')
+        xList = []
+        yList = []
+
+        for eachLine in dataList:
+                if len(eachLine) > 1:
+                        x, y = eachLine.split(' ')
+                        xList.append(int(x))
+                        yList.append(int(y))
+
+        a.clear()
+        a.title.set_text("Length " + str(len(xList)))
+        a.plot(xList,yList)
+        
 class Plot_Page(tk.Frame):
         def __init__(self, parent, controller):
                 tk.Frame.__init__(self, parent)
@@ -29,3 +65,14 @@ class Plot_Page(tk.Frame):
                 goToHomePage_button = ttk.Button(self, text="Home",
                                     command=lambda: controller.show_page("Home_Page"))
                 goToHomePage_button.pack(pady=1,padx=15, side = "left", expand = "no", anchor = "n")
+
+                canvas = FigureCanvasTkAgg(f, self)
+                canvas.show()
+                canvas.get_tk_widget().pack(side=tk.BOTTOM, fill=tk.BOTH, expand=True)
+
+                toolbar = NavigationToolbar2TkAgg(canvas, self)
+                toolbar.update()
+                canvas._tkcanvas.pack(side=tk.TOP, fill=tk.BOTH, expand=True)
+
+#------------------------------------------------------------------#
+#------------------------------------------------------------------#
