+++ ava.py
@@ -15,67 +15,7 @@ from tkinter import messagebox
 from tkinter import ttk
 
 import numpy as np
-
-# Import plotting libraries
-import matplotlib.pyplot as pl
-import matplotlib, sys
-import matplotlib.animation as animation
-from matplotlib import style
-
-# Plotting library canvas tool
-from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg, NavigationToolbar2TkAgg
-from matplotlib.figure import Figure
-matplotlib.use('TkAgg')
-
-f = Figure(figsize=(5,5), dpi=100)
-a = f.add_subplot(111)
-
-from sys import argv
-
-# Plot Page
-def animate(i):
-        filepath = "/home/pi/ava/"
-        os.chdir(filepath)
-        
-        pullData = open('DataPlotFile.txt','r').read()
-        dataList = pullData.split('\n')
-        xList = []
-        yList = []
-
-        for eachLine in dataList:
-                if len(eachLine) > 1:
-                        x, y = eachLine.split(' ')
-                        xList.append(int(x))
-                        yList.append(int(y))
-
-        a.clear()
-        a.title.set_text("Length " + str(len(xList)))
-        a.plot(xList,yList)
-        
-class Plot_Page(tk.Frame):
-        def __init__(self, parent, controller):
-                tk.Frame.__init__(self, parent)
-
-                # AVA app controller (app_data access)
-                self.controller = controller
-                
-                label = ttk.Label(self, text="Plot Page")
-                label.pack(pady=1,padx=1, side = "top", anchor = "n")
-
-                goToHomePage_button = ttk.Button(self, text="Home",
-                                    command=lambda: controller.show_page("Home_Page"))
-                goToHomePage_button.pack(pady=1,padx=15, side = "left", expand = "no", anchor = "n")
-
-                canvas = FigureCanvasTkAgg(f, self)
-                canvas.show()
-                canvas.get_tk_widget().pack(side=tk.BOTTOM, fill=tk.BOTH, expand=True)
-
-                toolbar = NavigationToolbar2TkAgg(canvas, self)
-                toolbar.update()
-                canvas._tkcanvas.pack(side=tk.TOP, fill=tk.BOTH, expand=True)
-
-#------------------------------------------------------------------#
-#------------------------------------------------------------------#                
+              
 class FullScreenApp(object):
     def __init__(self, master, **kwargs):
         self.master=master
@@ -118,6 +58,7 @@ class AVA(tk.Tk):
         from Test_Is_Running_Page import Test_Is_Running_Page
         from Save_Test_Page import Save_Test_Page
         from Results_Page import Results_Page
+        from Plot_Page import Plot_Page
 
         # Create frames for each of the app pages
         for F in (Home_Page, Configure_Test_Page, Plot_Page, New_Vehicle_Page, Test_Is_Running_Page, Save_Test_Page, Results_Page):
@@ -149,5 +90,7 @@ class AVA(tk.Tk):
 app = AVA()
 app.title("Automotive Vibration Analyzer")
 #fullscreen = FullScreenApp(app)
-animate = animation.FuncAnimation(f,animate,interval=1000)
+import Plot_Page
+import matplotlib.animation as animation
+animate = Plot_Page.animation.FuncAnimation(Plot_Page.f,Plot_Page.animate,interval=1000)
 app.mainloop()
